Authorization delete

+++ backend/src/main/java/com/back/p67260811/domain/post/post/controller/ApiV1PostController.java
@@ -63,7 +63,7 @@ public class ApiV1PostController {
     @Transactional
     public RsData<PostDto> write(
             @Valid @RequestBody PostWriteReqBody reqBody,
-            @RequestHeader("Authorization") @NotBlank @Size(min=30, max=50) String apiKey
+            @RequestHeader("Authorization") @NotBlank String apiKey
     ) {
 
         String authorization = apiKey.substring(7);
@@ -96,7 +96,7 @@ public class ApiV1PostController {
     public RsData<Void> modify(
             @PathVariable int id,
             @Valid @RequestBody PostModifyReqBody reqBody,
-            @RequestHeader("Authorization") @NotBlank @Size(min=30, max=50) String apiKey
+            @RequestHeader("Authorization") @NotBlank String apiKey
     ) {
         String authorization = apiKey.substring(7);
         Member actor = memberService.findByApiKey(authorization).orElseThrow(
@@ -119,8 +119,16 @@ public class ApiV1PostController {
 
     @DeleteMapping("/{id}")
     public RsData<Void> delete(
-            @PathVariable int id
+            @PathVariable int id,
+            @RequestHeader("Authorization") @NotBlank String apiKey
     ) {
+        String authorization = apiKey.replace("Bearer ", "");
+
+        Member actor = memberService.findByApiKey(authorization).orElseThrow(() -> new ServiceException("401-1", "API 키가 올바르지 않습니다."));
+        Post post = postService.findById(id).get();
+
+        if(!actor.equals(post.getAuthor())) throw new ServiceException("403-1", "삭제 권한이 없습니다.");
+
         postService.delete(id);
 
         return new RsData<>(
